Putted chat showcases data in separate file based on the different usecases
